# accounting/openapi/openapi_v3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Komunitin accounting API",
    "version": "0.0.1",
    "license": {
      "name": "AGPL-3.0-only"
    },
    "description": "API for the Komunitin payments system. This documentation is automatically generated from the serivce test suite."
  },
  "paths": {
...
    "/{code}/currency/settings": {
      "get": {
        "summary": "/{code}/currency/settings",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "currency"
        ]
      },
      "patch": {
        "summary": "/{code}/currency/settings",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "example": "TEST",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "object",
                    "properties": {
                      "attributes": {
                        "type": "object",
                        "properties": {
                          "defaultOnPaymentCreditLimit": {
                            "type": "number",
                            "example": 1300
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "currency-settings"
                        },
                        "id": {
                          "type": "string",
                          "example": "f3eeabe3-2772-49ae-be20-437a066f9a38"
                        },
                        "attributes": {
                          "type": "object",
                          "properties": {
                            "defaultAllowPayments": {
                              "type": "boolean",
                              "example": true
                            },
                            "enableExternalPayments": {
                              "type": "boolean",
                              "example": true
                            },
                            "defaultAllowTagPayments": {
                              "type": "boolean",
                              "example": false
                            },
                            "defaultInitialCreditLimit": {
                              "type": "number",
                              "example": 1000
                            },
                            "externalTraderCreditLimit": {
                              "type": "number",
                              "example": 1000
                            },
                            "defaultAcceptPaymentsAfter": {
                              "type": "number",
                              "example": 1209600
                            },
                            "defaultAllowPaymentRequests": {
                              "type": "boolean",
                              "example": true
                            },
                            "defaultOnPaymentCreditLimit": {
                              "type": "number",
                              "example": 1300
                            },
                            "defaultAllowExternalPayments": {
                              "type": "boolean",
                              "example": true
                            },
                            "enableExternalPaymentRequests": {
                              "type": "boolean",
                              "example": false
                            },
                            "defaultAcceptPaymentsWhitelist": {
                              "type": "array",
                              "items": {},
                              "example": [
                                null
                              ]
                            },
                            "defaultAllowTagPaymentRequests": {
                              "type": "boolean",
                              "example": false
                            },
                            "defaultAcceptPaymentsAutomatically": {
                              "type": "boolean",
                              "example": false
                            },
                            "defaultAllowExternalPaymentRequests": {
                              "type": "boolean",
                              "example": false
                            },
                            "defaultAcceptExternalPaymentsAutomatically": {
                              "type": "boolean",
                              "example": false
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "detail": {
                            "type": "string"
                          }
                        }
                      },
                      "example": [
                        {
                          "status": "400",
                          "code": "BadRequest",
                          "title": "Bad Request",
                          "detail": ""
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "detail": {
                            "type": "string"
                          }
                        }
                      },
                      "example": [
                        {
                          "status": "403",
                          "code": "Forbidden",
                          "title": "Forbidden",
                          "detail": "User not found in currency TES2"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "detail": {
                            "type": "string"
                          }
                        }
                      },
                      "example": [
                        {
                          "status": "404",
                          "code": "NotFound",
                          "title": "Not Found",
                          "detail": "Currency with code TES2 not found"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "detail": {
                            "type": "string"
                          }
                        }
                      },
                      "example": [
                        {
                          "status": "500",
                          "code": "NotImplemented",
                          "title": "Not Implemented",
                          "detail": "Change the external trader credit limit not implemented yet"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "currency"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
    "/{code}/trustlines/{id}": {
      "get": {
        "summary": "/{code}/trustlines/{id}",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "example": "EXTR",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "example": "d36f1173-ccfe-46a9-9995-c0c0d32800f6",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "trustlines"
                        },
                        "id": {
                          "type": "string",
                          "example": "79c07b98-5a60-43ca-91d7-58fda85e15e5"
                        },
                        "attributes": {
                          "type": "object",
                          "properties": {
                            "limit": {
                              "type": "number",
                              "example": 1000
                            },
                            "balance": {
                              "type": "number",
                              "example": 0
                            },
                            "created": {
                              "type": "string",
                              "example": "2024-10-17T10:32:43.097Z"
                            },
                            "updated": {
                              "type": "string",
                              "example": "2024-10-17T10:32:43.097Z"
                            }
                          }
                        },
                        "links": {
                          "type": "object",
                          "properties": {
                            "self": {
                              "type": "string",
                              "example": "http://localhost:2025/EXTR/trustlines/79c07b98-5a60-43ca-91d7-58fda85e15e5"
                            }
                          }
                        },
                        "relationships": {
                          "type": "object",
                          "properties": {
                            "currency": {
                              "type": "object",
                              "properties": {
                                "data": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "currencies"
                                    },
                                    "id": {
                                      "type": "string",
                                      "example": "476b2111-6778-4e76-b212-400dbdc1945c"
                                    }
                                  }
                                }
                              }
                            },
                            "trusted": {
                              "type": "object",
                              "properties": {
                                "data": {
                                  "type": "object",
                                  "properties": {
                                    "type": {
                                      "type": "string",
                                      "example": "currencies"
                                    },
                                    "id": {
                                      "type": "string",
                                      "example": "585bdb26-6e22-41ce-a677-85d8bf70a889"
                                    },
                                    "meta": {
                                      "type": "object",
                                      "properties": {
                                        "external": {
                                          "type": "boolean",
                                          "example": true
                                        },
                                        "href": {
                                          "type": "string",
                                          "example": "http://localhost:2025/TEST/currency"
                                        }
                                      }
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "included": {
                      "type": "array",
                      "items": {},
                      "example": [
                        null
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "trustlines"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "patch": {
        "summary": "/{code}/trustlines/{id}",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "trustlines"
        ]
      }
    },
    "/{code}/stats/volume": {
      "get": {
        "summary": "/{code}/stats/volume",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "example": "TEST",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "example": "2024-06-01",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "example": "2024-06-30",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "example": "P1M",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "stats"
                        },
                        "attributes": {
                          "type": "object",
                          "properties": {
                            "values": {
                              "type": "array",
                              "items": {
                                "type": "number"
                              },
                              "example": [
                                48970300
                              ]
                            },
                            "to": {
                              "type": "string",
                              "example": "2025-03-18T11:16:47.347Z"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "status": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "detail": {
                            "type": "string"
                          }
                        }
                      },
                      "example": [
                        {
                          "status": "500",
                          "code": "InternalError",
                          "title": "Internal Error",
                          "detail": "\nInvalid `prisma.$queryRaw()` invocation:\n\n\nRaw query failed. Code: `42883`. Message: `ERROR: function min(timestamp without time zone, timestamp without time zone) does not exist\nHINT: No function matches the given name and argument types. You might need to add explicit type casts.`"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://accounting.komunitin.org",
      "description": "Production server"
    },
    {
      "url": "https://accounting.demo.komunitin.org",
      "description": "Demo server"
    },
    {
      "url": "http://localhost:2025",
      "description": "Local development server"
    },
    {
      "url": "http://127.0.0.1:45881"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "authorization": {
        "name": "authorization",
        "in": "header",
        "type": "apiKey"
      }
    }
  }
}